+++ src/pages/Volunteers.jsx
@@ -4,7 +4,7 @@ import GeneralInfo from '../components/create-account/GeneralInfo';
 import WeeklyInfo from '../components/create-account/WeeklyInfo';
 import RolesAndSkills from '../components/create-account/RolesAndSkills';
 import DuiAndCrimHis from '../components/create-account/DuiAndCrimHis';
-import VolunteeringHistory from '../components/volunteer-profile-history/VolunteeringHistory';
+import VolunteeringHistory from './VolunteeringHistory';
 import 'antd/dist/antd.variable.min.css';
 
 function Volunteers() {
